% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVM.R
\name{CVM}
\alias{CVM}
\title{Two-sample Cramer-von Mises test for equal distributions}
\usage{
CVM(x, y)
}
\arguments{
\item{x}{: random samples from F}

\item{y}{: random samples from G}
}
\value{
p  : p-value of the test
}
\description{
A function used for Two-sample Cramer-von Mises testing for equal distributions
}
\examples{

attach(chickwts)
x <- sort(as.vector(weight[feed == "soybean"]))
y <- sort(as.vector(weight[feed == "linseed"]))
CVM(x,y)

}
