% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_beta.R
\name{gene_beta}
\alias{gene_beta}
\title{A beta(a,b) sampler using R}
\usage{
gene_beta(a, b, n)
}
\arguments{
\item{a, b}{the param of ditribution Beta(a,b)}

\item{n}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A beta(a,b) sampler using R
}
\examples{
\dontrun{
x=gene_beta(3,2,1000)
hist(x,prob=TRUE,main="1/beta(2,3)*x^2*(1-x)")
y=seq(0,1,0.01)
lines(y,1/beta(3,2)*y^2*(1-y))
}
}
