% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_chi.R
\name{my_chi}
\alias{my_chi}
\title{computes the chi-square test statistic for two vectors}
\usage{
my_chi(x, y)
}
\arguments{
\item{x, y}{two vectors}
}
\value{
chi-square test statistic
}
\description{
a faster version of chisq.test() that only computes the chi-square test statistic when the input is two numeric vectors with no missing values
}
\examples{
\dontrun{
x<-runif(10)
y<-runif(10)
my_chi(x,y)
}
}
